backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
django-filter>=23.5
gunicorn>=21.2.0
whitenoise>=6.6.0
